+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Center alignment
+
 ## [0.2.4] - 2019-11-13
 
 ## [0.2.3] - 2019-11-06